move products logic to seperate controller file
